rename dependencies to definitions
